refactor the localPT and eliminate the cycle dependence
